+++ server/api/index.js
@@ -2,7 +2,8 @@ const router = require('express').Router()
 module.exports = router
 
 router.use('/users', require('./users'))
-router.use('/cats', require('./cats'))
+router.use('/foods', require('./foods'))
+router.use('/images', require('./images'))
 
 router.use((req, res, next) => {
   const error = new Error('Not Found')
